Chore: Remove More Dead Code Found by Vulture in api/parsing
get_field_type()/DB_NAME_TO_FIELD_TYPE, FORMAT_NAME_TO_CODE, and a
duplicate COMPARISON_OPERATORS (the underscored _COMPARISON_OPERATORS
is the one actually used) were all orphaned.

extract_frame_data_from_raw_card() moved from card_query_nodes.py to
card_processing.py, its actual home, and preprocess_card() now calls
it instead of duplicating the same nine lines inline.
